updated product code
