+++ frontend/src/components/products/Products.module.css
@@ -13,68 +13,3 @@
   padding: 1rem;
   margin: 0;
 }
-
-.product {
-  position: relative;
-  z-index: 0;
-  display: flex;
-  flex-direction: column;
-  width: 250px;
-  color: #000000;
-  background: #f4f0f0;
-  border: 1px solid #9CAFAA;
-  cursor: pointer;
-}
-
-.product:hover {
-  z-index: 1;
-}
-
-.text {
-  border-top: 1px solid #9cafaa;
-  margin: 0;
-  padding: 0.5rem;
-  height: 3rem;
-}
-
-.text p {
-  max-height: 1rem;
-  margin: 0;
-  font-size: smaller;
-  text-align: center;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.username {
-  margin: 0;
-  padding: 0.25rem;
-  font-size: smaller;
-  background: linear-gradient(to right, #232526, #414345);
-  color: #f4f0f0;
-  border-right: 15px solid #414345;
-  overflow-x: auto;
-  text-wrap: nowrap;
-}
-
-.product img {
-  object-fit: cover;
-  width: 100%;
-  height: 250px;
-}
-
-@media (min-width: 320px) and (max-width: 480px) {
-  .product {
-    width: 120px;
-  }
-  .product img {
-    height: 120px;
-  }
-  .text {
-    height: 2.5rem;
-  }
-  .text p, .username {
-    font-size: x-small;
-  }
-}
