+++ frontend/src/components/products/Products.tsx
@@ -1,17 +1,16 @@
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { AnimatePresence, LayoutGroup, motion } from 'framer-motion';
 import Product from '@/models/Product';
 import css from './Products.module.css';
+import ProductItem from './Product';
 
 export default function Products({
   products,
   expanded,
 }: {
-  products: Product[];
+   products: Product[];
   expanded?: boolean | null;
 }) {
-  const navigate = useNavigate();
   const [searchTerm, setSearchTerm] = useState('');
 
   const filteredProducts = products?.filter((product) =>
@@ -25,53 +24,32 @@ export default function Products({
   return (
     <>
       <input
-          className={css.search}
+          className={css['search']}
               value={searchTerm}
            onChange={changeHandler}
         placeholder='Search...'
       />
       <LayoutGroup>
         <motion.ul
-          className={css.products}
-          layout
-          initial='hidden'
-          animate='visible'
+           className={css['products']}
+              layout
+             initial='hidden'
+             animate='visible'
           transition={{ staggerChildren: 0.15 }}
         >
           <AnimatePresence>
             {filteredProducts.length > 0 ? (
-              filteredProducts.map(({ _id, title, description, price, imageUrl, userId }) => {
-                return (
-                  <motion.li
-                    className={css.product}
-                    key={_id}
-                    layout
-                    variants={{
-                      hidden: { opacity: 0, y: -50, height: 0 },
-                      visible: { opacity: 1, y: 0, height: 'auto' },
-                    }}
-                    exit={{ opacity: 0, scale: 0, height: 0 }}
-                    whileHover={{ borderColor: '#000', y: -5 }}
-                    transition={{ type: 'tween', ease: 'easeInOut', duration: 0.2 }}
-                    onClick={() => navigate('/market/' + _id)}
-                  >
-                    <img src={`http://localhost:3000/${imageUrl}`} alt='product' />
-                    {expanded && <p className={css.username}>Posted by {userId.username}</p>}
-                    <div className={css.text}>
-                      <p>
-                        ${price.toFixed(2)} ○ {title}
-                      </p>
-                      <p style={{ color: '#7a7676' }}>{description}</p>
-                    </div>
-                  </motion.li>
-                );
-              })
+              filteredProducts.map((product) => (
+                <ProductItem key={product._id} product={product} expanded={expanded} />
+              ))
             ) : (
               <motion.p
-                layout
-                initial={{ opacity: 0, scale: 0, height: 100 }}
-                animate={{ opacity: 1, scale: 1, height: 'auto' }}
-                   exit={{ opacity: 0, scale: 0, height: 100 }}
+                       key='fallback'
+                    layout
+                   initial={{ opacity: 0, scale: 0, x: -350 }}
+                   animate={{ opacity: 1, scale: 1, x: 0 }}
+                      exit={{ opacity: 0, scale: 0 }}
+                transition={{ duration: 0.5 }}
               >
                 Nothing Found
               </motion.p>
